+++ lib/laflem/collections/python/__init__.py
@@ -3,7 +3,11 @@ Define the Python collection.
 """
 from laflem.collections.base import BaseCollection
 
-from .virtualenv import VirtualenvCleanupModule, VirtualenvCreateModule
+from .virtualenv import (
+    VirtualenvCleanupModule,
+    VirtualenvCreateModule,
+    VirtualenvListModule,
+)
 
 
 class PythonCollection(BaseCollection):
@@ -14,6 +18,7 @@ class PythonCollection(BaseCollection):
     name = "py"
     description = "The Python collection."
     modules = {
-        "venv-cleanup": VirtualenvCleanupModule,
         "venv-create": VirtualenvCreateModule,
+        "venv-list": VirtualenvListModule,
+        "venv-cleanup": VirtualenvCleanupModule,
     }