+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [Unreleased]
+
+### Features
+
+- Add `venv-list` module to `python` collection
+
+
 ## v0.3.0
 
 ### Features